Add steps & omit the loadbalancer & running via UI
* We can omit the service with the Loadbalancer, and simply expose the port to our
localhost only. This should add an additional guarantee in the event they accidentally
deploy it to a live cluster.
* Added additional warnings.
* We can’t really reference work.withpixie.ai because it’s not the OSS offering, I got
around that by adding instructions to run the Pixie CLI px/mysql_data command. It prints
out a URL afterward (like all Pixie CLI commands) that they can use to navigate to the
Live UI. The URL will work whether or not they are self hosted or using hosted pixie.
* Added a few steps that were previously implicit.

Signed-off-by: Hannah Stepanek <hstepanek@newrelic.com>