React Tables demo.
Adds both the demo (actually five demos), and GitHub Pages for this
repository so that the demos can go live.

For upcoming blog post [Tables are Hard: Part 2,
Basics](https://blog.px.dev/tables-are-hard-2).

As part of merging this diff, we'll need to enable GitHub Pages for this
repository. The default directory (root) will do the trick.

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>